Creating replacement collection factory
